requirements.txt
# Web Framework & Server
fastapi==0.104.1
uvicorn==0.24.0

# Database ORM
sqlalchemy==2.0.23

# Data Collection Services
google-api-python-client==2.111.0
pytrends==4.9.2
requests==2.31.0
httpx==0.25.2

# Configuration & Utilities
python-dotenv==1.0.0
apscheduler==3.10.4
pandas==2.1.3